+++ tests/nanomodbus_tests.c
@@ -5,7 +5,7 @@
 #include <string.h>
 
 
-int read_empty(uint8_t* b, uint32_t count, int32_t timeout, void* arg) {
+int32_t read_empty(uint8_t* b, uint16_t count, int32_t timeout, void* arg) {
     UNUSED_PARAM(b);
     UNUSED_PARAM(count);
     UNUSED_PARAM(timeout);
@@ -14,7 +14,7 @@ int read_empty(uint8_t* b, uint32_t count, int32_t timeout, void* arg) {
 }
 
 
-int write_empty(const uint8_t* b, uint32_t count, int32_t timeout, void* arg) {
+int32_t write_empty(const uint8_t* b, uint16_t count, int32_t timeout, void* arg) {
     UNUSED_PARAM(b);
     UNUSED_PARAM(count);
     UNUSED_PARAM(timeout);
@@ -73,7 +73,7 @@ void test_server_create(nmbs_transport transport) {
 }
 
 
-int read_timeout(uint8_t* buf, uint32_t count, int32_t timeout, void* arg) {
+int32_t read_timeout(uint8_t* buf, uint16_t count, int32_t timeout, void* arg) {
     UNUSED_PARAM(buf);
     UNUSED_PARAM(count);
     UNUSED_PARAM(arg);
@@ -82,7 +82,7 @@ int read_timeout(uint8_t* buf, uint32_t count, int32_t timeout, void* arg) {
 }
 
 // Timeouts on the second read
-int read_timeout_second(uint8_t* buf, uint32_t count, int32_t timeout, void* arg) {
+int32_t read_timeout_second(uint8_t* buf, uint16_t count, int32_t timeout, void* arg) {
     UNUSED_PARAM(count);
     UNUSED_PARAM(arg);
